==> 4. Решение рядов с помощью классических алгоритмов/lesson4/step4_linreg_CV.ipynb <==
%matplotlib inline
import matplotlib.pyplot as plt

import numpy as np
import pandas as pd 
import seaborn as sns


from sklearn.metrics import mean_squared_error, r2_score
from sklearn.linear_model import LinearRegression, Ridge, Lasso
from sklearn.preprocessing import StandardScaler
import os

train = pd.read_csv('train_prepared.csv', index_col=[0], parse_dates=[0])
#test = pd.read_csv('test_prepared.csv', index_col=[0], parse_dates=[0])

train.head()

Сделаем кроссвалидацию и сохраним оценки

lr = LinearRegression()
sc = StandardScaler()
X = train.drop('Passengers', 1).values
y = train['Passengers'].values
scaler = sc.fit(X)
X = sc.transform(X)

from sklearn.model_selection import TimeSeriesSplit

tscv = TimeSeriesSplit(10)

CV_scores_lr = []
for train_index, test_index in tscv.split(X):
    print("TRAIN:", train_index, "TEST:", test_index)
    X_train, X_test = X[train_index], X[test_index]
    y_train, y_test = y[train_index], y[test_index]
    lr = LinearRegression()
    lr.fit(X_train, y_train)
    preds = lr.predict(X_test)
    CV_scores_lr.append(r2_score(y_test, preds))

CV_scores_lr

Сохраним обученную модель

import dill

lr = LinearRegression()
lr.fit(X, y)
with open("models/LR.model", "wb") as f:
    dill.dump(lr, f)

### Ridge regression

tscv = TimeSeriesSplit(10)

CV_scores_lr_l2 = []
for train_index, test_index in tscv.split(X):
    print("TRAIN:", train_index, "TEST:", test_index)
    X_train, X_test = X[train_index], X[test_index]
    y_train, y_test = y[train_index], y[test_index]
    lr = Ridge()
    lr.fit(X_train, y_train)
    preds = lr.predict(X_test)
    CV_scores_lr_l2.append(r2_score(y_test, preds))

lr = Ridge()
lr.fit(X, y)
with open("models/LR_l2.model", "wb") as f:
    dill.dump(lr, f)

### Lasso regression

tscv = TimeSeriesSplit(10)

CV_scores_lr_l1 = []
for train_index, test_index in tscv.split(X):
    print("TRAIN:", train_index, "TEST:", test_index)
    X_train, X_test = X[train_index], X[test_index]
    y_train, y_test = y[train_index], y[test_index]
    lr = Lasso()
    lr.fit(X_train, y_train)
    preds = lr.predict(X_test)
    CV_scores_lr_l1.append(r2_score(y_test, preds))
    
lr = Lasso()
lr.fit(X, y)
with open("models/LR_l1.model", "wb") as f:
    dill.dump(lr, f)

pd.DataFrame({'LR': CV_scores_lr, 
              'LR_l2': CV_scores_lr_l1,
              'LR_l1': CV_scores_lr_l2}).to_csv("./models/CV_LR.csv")

pd.DataFrame({'LR': CV_scores_lr, 
              'LR_l2': CV_scores_lr_l1,
              'LR_l1': CV_scores_lr_l2})

## ПЕРЕРЫВ до 21:00
